refactor: split assistant-runtime into tool-definitions, tool-executors, prompt-analysis
Extract prompt analysis, tool definition building, and tool execution
into dedicated modules from assistant-runtime.ts, keeping only the
main turn orchestration logic in the original file.
